EES-7131 - Move StatisticalReleasesGuidanceModal.tsx to components directory. Remove unnecessary sections.
